requirements.txt
pyautogen
openai
faiss-cpu
numpy
pandas
python-dotenv